+++ executor/simple_test.go
@@ -21,7 +21,6 @@ import (
 	"testing"
 
 	"github.com/pingcap/tidb/config"
-	"github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/parser/auth"
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tidb/server"
@@ -131,220 +130,4 @@ func TestUserAttributes(t *testing.T) {
 	rootTK.MustQuery("select user_attributes from mysql.user where user = 'usr1'").Check(testkit.Rows(`{"metadata": {"comment": "comment1"}, "resource_group": "default"}`))
 	rootTK.MustExec("alter user usr1 resource group 'rg1'")
 	rootTK.MustQuery("select user_attributes from mysql.user where user = 'usr1'").Check(testkit.Rows(`{"metadata": {"comment": "comment1"}, "resource_group": "rg1"}`))
-}
-
-func TestValidatePassword(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	subtk := testkit.NewTestKit(t, store)
-	err := tk.Session().Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil)
-	require.NoError(t, err)
-	tk.MustExec("CREATE USER ''@'localhost'")
-	tk.MustExec("GRANT ALL PRIVILEGES ON mysql.* TO ''@'localhost';")
-	err = subtk.Session().Auth(&auth.UserIdentity{Hostname: "localhost"}, nil, nil)
-	require.NoError(t, err)
-
-	authPlugins := []string{mysql.AuthNativePassword, mysql.AuthCachingSha2Password, mysql.AuthTiDBSM3Password}
-	tk.MustQuery("SELECT @@global.validate_password.enable").Check(testkit.Rows("0"))
-	tk.MustExec("SET GLOBAL validate_password.enable = 1")
-	tk.MustQuery("SELECT @@global.validate_password.enable").Check(testkit.Rows("1"))
-
-	for _, authPlugin := range authPlugins {
-		tk.MustExec("DROP USER IF EXISTS testuser")
-		tk.MustExec(fmt.Sprintf("CREATE USER testuser IDENTIFIED WITH %s BY '!Abc12345678'", authPlugin))
-
-		tk.MustExec("SET GLOBAL validate_password.policy = 'LOW'")
-		// check user name
-		tk.MustQuery("SELECT @@global.validate_password.check_user_name").Check(testkit.Rows("1"))
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY '!Abcdroot1234'", "Password Contains User Name")
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY '!Abcdtoor1234'", "Password Contains Reversed User Name")
-		tk.MustExec("SET PASSWORD FOR 'testuser' = 'testuser'") // password the same as the user name, but run by root
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY 'testuser'")
-		tk.MustExec("SET GLOBAL validate_password.check_user_name = 0")
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY '!Abcdroot1234'")
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY '!Abcdtoor1234'")
-		tk.MustExec("SET GLOBAL validate_password.check_user_name = 1")
-
-		// LOW: Length
-		tk.MustExec("SET GLOBAL validate_password.length = 8")
-		tk.MustQuery("SELECT @@global.validate_password.length").Check(testkit.Rows("8"))
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY '1234567'", "Require Password Length: 8")
-		tk.MustExec("SET GLOBAL validate_password.length = 12")
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY '!Abcdefg123'", "Require Password Length: 12")
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY '!Abcdefg1234'")
-		tk.MustExec("SET GLOBAL validate_password.length = 8")
-
-		// MEDIUM: Length; numeric, lowercase/uppercase, and special characters
-		tk.MustExec("SET GLOBAL validate_password.policy = 'MEDIUM'")
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY '!Abc1234567'")
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY '!ABC1234567'", "Require Password Lowercase Count: 1")
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY '!abc1234567'", "Require Password Uppercase Count: 1")
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY '!ABCDabcd'", "Require Password Digit Count: 1")
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY 'Abc1234567'", "Require Password Non-alphanumeric Count: 1")
-		tk.MustExec("SET GLOBAL validate_password.special_char_count = 0")
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY 'Abc1234567'")
-		tk.MustExec("SET GLOBAL validate_password.special_char_count = 1")
-		tk.MustExec("SET GLOBAL validate_password.length = 3")
-		tk.MustQuery("SELECT @@GLOBAL.validate_password.length").Check(testkit.Rows("4"))
-
-		// STRONG: Length; numeric, lowercase/uppercase, and special characters; dictionary file
-		tk.MustExec("SET GLOBAL validate_password.policy = 'STRONG'")
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY '!Abc1234567'")
-		tk.MustExec(fmt.Sprintf("SET GLOBAL validate_password.dictionary = '%s'", "1234;5678"))
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY '!Abc123567'")
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY '!Abc43218765'")
-		tk.MustContainErrMsg("ALTER USER testuser IDENTIFIED BY '!Abc1234567'", "Password contains word in the dictionary")
-		tk.MustExec("SET GLOBAL validate_password.dictionary = ''")
-		tk.MustExec("ALTER USER testuser IDENTIFIED BY '!Abc1234567'")
-
-		// "IDENTIFIED AS 'xxx'" is not affected by validation
-		tk.MustExec(fmt.Sprintf("ALTER USER testuser IDENTIFIED WITH '%s' AS ''", authPlugin))
-	}
-	tk.MustContainErrMsg("CREATE USER 'testuser1'@'localhost'", "Your password does not satisfy the current policy requirements")
-	tk.MustContainErrMsg("CREATE USER 'testuser1'@'localhost' IDENTIFIED WITH 'caching_sha2_password'", "Your password does not satisfy the current policy requirements")
-	tk.MustContainErrMsg("CREATE USER 'testuser1'@'localhost' IDENTIFIED WITH 'caching_sha2_password' AS ''", "Your password does not satisfy the current policy requirements")
-
-	// if the username is '', all password can pass the check_user_name
-	subtk.MustQuery("SELECT user(), current_user()").Check(testkit.Rows("@localhost @localhost"))
-	subtk.MustQuery("SELECT @@global.validate_password.check_user_name").Check(testkit.Rows("1"))
-	subtk.MustQuery("SELECT @@global.validate_password.enable").Check(testkit.Rows("1"))
-	tk.MustExec("SET GLOBAL validate_password.number_count = 0")
-	tk.MustExec("SET GLOBAL validate_password.special_char_count = 0")
-	tk.MustExec("SET GLOBAL validate_password.mixed_case_count = 0")
-	tk.MustExec("SET GLOBAL validate_password.length = 0")
-	subtk.MustExec("ALTER USER ''@'localhost' IDENTIFIED BY ''")
-	subtk.MustExec("ALTER USER ''@'localhost' IDENTIFIED BY 'abcd'")
-
-	// CREATE ROLE is not affected by password validation
-	tk.MustExec("SET GLOBAL validate_password.enable = 1")
-	tk.MustExec("SET GLOBAL validate_password.number_count = default")
-	tk.MustExec("SET GLOBAL validate_password.special_char_count = default")
-	tk.MustExec("SET GLOBAL validate_password.mixed_case_count = default")
-	tk.MustExec("SET GLOBAL validate_password.length = default")
-	tk.MustExec("CREATE ROLE role1")
-}
-
-func expectedPasswordExpiration(t *testing.T, tk *testkit.TestKit, testuser, expired string, lifetime string) {
-	res := tk.MustQuery(fmt.Sprintf("SELECT password_expired, password_last_changed, password_lifetime FROM mysql.user WHERE user = '%s'", testuser))
-	rows := res.Rows()
-	require.NotEmpty(t, rows)
-	row := rows[0]
-	require.Equal(t, 3, len(row))
-	require.Equal(t, expired, row[0].(string), testuser)
-	require.True(t, len(row[1].(string)) > 0, testuser)
-	require.Equal(t, lifetime, row[2].(string), testuser)
-}
-
-func TestPasswordExpiration(t *testing.T) {
-	store, _ := testkit.CreateMockStoreAndDomain(t)
-	tk := testkit.NewTestKit(t, store)
-
-	// CREATE USER
-	tk.MustExec(`CREATE USER testuser`)
-	expectedPasswordExpiration(t, tk, "testuser", "N", "<nil>")
-	tk.MustExec(`CREATE USER testuser1 PASSWORD EXPIRE`)
-	expectedPasswordExpiration(t, tk, "testuser1", "Y", "<nil>")
-	tk.MustExec(`CREATE USER testuser2 PASSWORD EXPIRE DEFAULT`)
-	expectedPasswordExpiration(t, tk, "testuser2", "N", "<nil>")
-	tk.MustExec(`CREATE USER testuser3 PASSWORD EXPIRE NEVER`)
-	expectedPasswordExpiration(t, tk, "testuser3", "N", "0")
-	tk.MustExec(`CREATE USER testuser4 PASSWORD EXPIRE INTERVAL 3 DAY`)
-	expectedPasswordExpiration(t, tk, "testuser4", "N", "3")
-	tk.MustExec(`CREATE ROLE role1`)
-	expectedPasswordExpiration(t, tk, "role1", "Y", "<nil>")
-
-	// ALTER USER
-	testcases := []struct {
-		user    string
-		expired string
-	}{
-		{"testuser", "N"},
-		{"testuser1", "Y"},
-		{"testuser2", "N"},
-		{"testuser3", "N"},
-		{"testuser4", "N"},
-		{"role1", "Y"},
-	}
-	for _, testcase := range testcases {
-		tk.MustExec(fmt.Sprintf("ALTER USER %s PASSWORD EXPIRE NEVER", testcase.user))
-		expectedPasswordExpiration(t, tk, testcase.user, testcase.expired, "0")
-		tk.MustExec(fmt.Sprintf("ALTER USER %s PASSWORD EXPIRE DEFAULT", testcase.user))
-		expectedPasswordExpiration(t, tk, testcase.user, testcase.expired, "<nil>")
-		tk.MustExec(fmt.Sprintf("ALTER USER %s PASSWORD EXPIRE INTERVAL 3 DAY", testcase.user))
-		expectedPasswordExpiration(t, tk, testcase.user, testcase.expired, "3")
-		tk.MustExec(fmt.Sprintf("ALTER USER %s PASSWORD EXPIRE", testcase.user))
-		expectedPasswordExpiration(t, tk, testcase.user, "Y", "3")
-		tk.MustExec(fmt.Sprintf("ALTER USER %s IDENTIFIED BY '' PASSWORD EXPIRE", testcase.user))
-		expectedPasswordExpiration(t, tk, testcase.user, "Y", "3")
-		tk.MustExec(fmt.Sprintf("ALTER USER %s IDENTIFIED WITH 'mysql_native_password' AS ''", testcase.user))
-		expectedPasswordExpiration(t, tk, testcase.user, "N", "3")
-		tk.MustExec(fmt.Sprintf("ALTER USER %s IDENTIFIED BY ''", testcase.user))
-		expectedPasswordExpiration(t, tk, testcase.user, "N", "3")
-	}
-
-	// SET PASSWORD
-	tk.MustExec("ALTER USER testuser PASSWORD EXPIRE")
-	expectedPasswordExpiration(t, tk, "testuser", "Y", "3")
-	tk.MustExec("SET PASSWORD FOR testuser = '1234'")
-	expectedPasswordExpiration(t, tk, "testuser", "N", "3")
-
-	tk.MustGetErrCode(`CREATE USER ''@localhost IDENTIFIED BY 'pass' PASSWORD EXPIRE`, mysql.ErrPasswordExpireAnonymousUser)
-	tk.MustExec(`CREATE USER ''@localhost IDENTIFIED BY 'pass'`)
-	tk.MustGetErrCode(`ALTER USER ''@localhost PASSWORD EXPIRE`, mysql.ErrPasswordExpireAnonymousUser)
-
-	// different cleartext authentication plugin
-	for _, authplugin := range []string{mysql.AuthNativePassword, mysql.AuthCachingSha2Password, mysql.AuthTiDBSM3Password} {
-		tk.MustExec("DROP USER IF EXISTS 'u1'@'localhost'")
-		tk.MustExec(fmt.Sprintf("CREATE USER 'u1'@'localhost' IDENTIFIED WITH '%s'", authplugin))
-		tk.MustExec("ALTER USER 'u1'@'localhost' IDENTIFIED BY 'pass'")
-		tk.MustExec("ALTER USER 'u1'@'localhost' PASSWORD EXPIRE")
-		tk.MustQuery("SELECT password_expired FROM mysql.user WHERE user = 'u1'").Check(testkit.Rows("Y"))
-	}
-}
-
-// Test cases that related to PASSWORD VALIDATION, PASSWORD EXPIRATION, PASSWORD REUSE POLICY, and PASSWORD FAILED-LOGIN TRACK.
-func TestPasswordManagement(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("SET GLOBAL validate_password.enable = 1")
-	defer tk.MustExec("SET GLOBAL validate_password.enable = 0")
-
-	// PASSWORD VALIDATION can work with user-specified PASSWORD REUSE POLICY.
-	tk.MustExec("CREATE USER u1 IDENTIFIED BY '!Abc1234' password history 1")
-	tk.MustGetErrCode("ALTER USER u1 IDENTIFIED BY '!Abc1234'", errno.ErrExistsInHistoryPassword)
-	tk.MustGetErrCode("ALTER USER u1 IDENTIFIED BY '!abc1234'", errno.ErrNotValidPassword)
-
-	// PASSWORD VALIDATION can work with global PASSWORD REUSE POLICY.
-	tk.MustExec("SET GLOBAL password_history = 1")
-	tk.MustExec("DROP USER u1")
-	tk.MustExec("CREATE USER u1 IDENTIFIED BY '!Abc1234'")
-	tk.MustGetErrCode("ALTER USER u1 IDENTIFIED BY '!Abc1234'", errno.ErrExistsInHistoryPassword)
-	tk.MustGetErrCode("ALTER USER u1 IDENTIFIED BY '!abc1234'", errno.ErrNotValidPassword)
-
-	// PASSWORD EXPIRATION can work with ACCOUNT LOCK.
-	// PASSWORD EXPIRE NEVER and ACCOUNT UNLOCK take effect.
-	tk.MustExec(`ALTER USER u1 ACCOUNT LOCK PASSWORD EXPIRE NEVER PASSWORD EXPIRE NEVER ACCOUNT UNLOCK ACCOUNT LOCK ACCOUNT LOCK ACCOUNT UNLOCK;`)
-	tk.MustQuery(`SELECT password_expired, password_lifetime, account_locked FROM mysql.user WHERE USER='u1';`).Check(
-		testkit.Rows("N 0 N"))
-
-	// PASSWORD EXPIRATION can work with PASSWORD REUSE POLICY
-	tk.MustExec(`create user u2 identified by '!Abc1234' password expire password reuse interval default password expire never password
-		reuse interval 3 day password history 5 password history default password expire default`)
-	tk.MustQuery(`select password_expired, password_lifetime, password_reuse_history, password_reuse_time from mysql.user where user = 'u2'`).Check(
-		testkit.Rows("N <nil> <nil> 3"))
-	tk.MustExec(`alter user u2 password expire default password reuse interval 3 day password history default
-		password expire never password expire interval 5 day password reuse interval default password expire password history 5`)
-	tk.MustQuery(`select password_expired, password_lifetime, password_reuse_history, password_reuse_time from mysql.user where user = 'u2'`).Check(
-		testkit.Rows("Y <nil> 5 <nil>"))
-	tk.MustExec(`alter user u2 identified by '!Abc12345'`)
-	tk.MustQuery(`select password_expired, password_lifetime, password_reuse_history, password_reuse_time from mysql.user where user = 'u2'`).Check(
-		testkit.Rows("N <nil> 5 <nil>"))
-
-	// PASSWORD FAILED-LOGIN TRACK can work with USER COMMENT and USER ATTRIBUTE
-	tk.MustExec(`CREATE USER u3 IDENTIFIED BY '!Abc12345' FAILED_LOGIN_ATTEMPTS 4 PASSWORD_LOCK_TIME 3 COMMENT 'Some statements to test create user'`)
-	tk.MustQuery(`select user_attributes->>"$.metadata" from mysql.user where user = 'u3'`).Check(testkit.Rows(`{"comment": "Some statements to test create user"}`))
-	tk.MustQuery(`select user_attributes->>"$.Password_locking" from mysql.user where user = 'u3'`).Check(testkit.Rows(`{"failed_login_attempts": 4, "password_lock_time_days": 3}`))
-	tk.MustExec(`ALTER USER u3 FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME unbounded FAILED_LOGIN_ATTEMPTS 5 PASSWORD_LOCK_TIME 5 ATTRIBUTE '{"name": "John", "age": 19}'`)
-	tk.MustQuery(`select user_attributes->>"$.metadata" from mysql.user where user = 'u3'`).Check(testkit.Rows(`{"age": 19, "comment": "Some statements to test create user", "name": "John"}`))
-	tk.MustQuery(`select user_attributes->>"$.Password_locking" from mysql.user where user = 'u3'`).Check(testkit.Rows(`{"failed_login_attempts": 5, "password_lock_time_days": 5}`))
 }
