+++ executor/simpletest/simple_test.go
@@ -15,18 +15,12 @@
 package simpletest
 
 import (
-	"bytes"
 	"context"
-	"crypto/sha1"
-	"encoding/json"
-	"fmt"
 	"strconv"
-	"strings"
 	"testing"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/config"
-	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/executor"
 	"github.com/pingcap/tidb/parser/auth"
 	"github.com/pingcap/tidb/parser/model"
@@ -35,11 +29,9 @@ import (
 	"github.com/pingcap/tidb/planner/core"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx"
-	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/statistics/handle"
 	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/pingcap/tidb/testkit"
-	"github.com/pingcap/tidb/util/sqlexec"
 	"github.com/stretchr/testify/require"
 	"go.opencensus.io/stats/view"
 )
@@ -888,109 +880,6 @@ func TestUser(t *testing.T) {
 	tk.MustQuery("select user from mysql.user where user='engineering' and host = 'us'").Check(testkit.Rows())
 }
 
-func TestFailedLoginTracking(t *testing.T) {
-	type passwordLocking struct {
-		FailedLoginAttempts   int64  `json:"failed_login_attempts"`
-		PasswordLockTimeDays  int64  `json:"password_lock_time_days"`
-		AutoAccountLocked     string `json:"auto_account_locked"`
-		FailedLoginCount      int64  `json:"failed_login_count"`
-		AutoLockedLastChanged string `json:"auto_locked_last_changed"`
-	}
-	type metadata struct {
-		Comment string `json:"comment"`
-	}
-	type userAttributes struct {
-		PasswordLocking passwordLocking `json:"Password_locking"`
-		Metadata        metadata        `json:"metadata"`
-	}
-
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-
-	createUserTestCases := []struct {
-		sql    string
-		rsJSON string
-		user   string
-	}{
-		{"CREATE USER 'u1'@'localhost' IDENTIFIED BY 'password' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 3;",
-			"{\"failed_login_attempts\": 3, \"password_lock_time_days\": 3}", "u1"},
-		{"CREATE USER 'u2'@'localhost' IDENTIFIED BY 'password' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME UNBOUNDED;",
-			"{\"failed_login_attempts\": 3, \"password_lock_time_days\": -1}", "u2"},
-		{"CREATE USER 'u3'@'localhost' IDENTIFIED BY 'password' FAILED_LOGIN_ATTEMPTS 3;",
-			"{\"failed_login_attempts\": 3, \"password_lock_time_days\": 0}", "u3"},
-		{"CREATE USER 'u4'@'localhost' IDENTIFIED BY 'password' PASSWORD_LOCK_TIME 3;",
-			"{\"failed_login_attempts\": 0, \"password_lock_time_days\": 3}", "u4"},
-		{"CREATE USER 'u5'@'localhost' IDENTIFIED BY 'password' PASSWORD_LOCK_TIME UNBOUNDED;",
-			"{\"failed_login_attempts\": 0, \"password_lock_time_days\": -1}", "u5"},
-	}
-	for _, tc := range createUserTestCases {
-		tk.MustExec(tc.sql)
-		sql := fmt.Sprintf("SELECT user_attributes->>\"$.Password_locking\" from mysql.user WHERE USER = '%s' AND HOST = 'localhost' for update", tc.user)
-		tk.MustQuery(sql).Check(testkit.Rows(tc.rsJSON))
-	}
-
-	alterUserTestCases := []struct {
-		sql                  string
-		user                 string
-		failedLoginAttempts  int64
-		passwordLockTimeDays int64
-		failedLoginCount     int64
-		comment              string
-	}{
-		{"ALTER USER 'u1'@'localhost' FAILED_LOGIN_ATTEMPTS 4 PASSWORD_LOCK_TIME 6;", "u1",
-			4, 6, 0, ""},
-		{"ALTER USER 'u2'@'localhost' FAILED_LOGIN_ATTEMPTS 4 PASSWORD_LOCK_TIME UNBOUNDED;",
-			"u2", 4, -1, 0, ""},
-		{"ALTER USER 'u3'@'localhost' PASSWORD_LOCK_TIME 6;",
-			"u3", 3, 6, 0, ""},
-		{"ALTER USER 'u4'@'localhost' FAILED_LOGIN_ATTEMPTS 4;",
-			"u4", 4, 3, 0, ""},
-		{"ALTER USER 'u4'@'localhost' PASSWORD_LOCK_TIME UNBOUNDED;",
-			"u4", 4, -1, 0, ""},
-		{"ALTER USER 'u5'@'localhost' ACCOUNT UNLOCK FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 6;",
-			"u5", 3, 6, 0, ""},
-		{"ALTER USER 'u5'@'localhost' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 6 COMMENT 'Something';",
-			"u5", 3, 6, 0, "Something"},
-	}
-	for _, tc := range alterUserTestCases {
-		tk.MustExec(tc.sql)
-		sql := fmt.Sprintf("SELECT user_attributes from mysql.user WHERE USER = '%s' AND HOST = 'localhost' for update", tc.user)
-		rs := tk.MustQuery(sql)
-		buf := bytes.NewBufferString("")
-		for _, row := range rs.Rows() {
-			_, err := fmt.Fprintf(buf, "%s\n", row)
-			require.NoError(t, err)
-		}
-		str := buf.String()
-		var ua []userAttributes
-		err := json.Unmarshal([]byte(str), &ua)
-		require.NoError(t, err)
-		require.Equal(t, tc.failedLoginAttempts, ua[0].PasswordLocking.FailedLoginAttempts, tc.sql, str)
-		require.Equal(t, tc.passwordLockTimeDays, ua[0].PasswordLocking.PasswordLockTimeDays, tc.sql, str)
-		require.Equal(t, tc.failedLoginCount, ua[0].PasswordLocking.FailedLoginCount, tc.sql, str)
-		require.Equal(t, tc.comment, ua[0].Metadata.Comment, tc.sql, str)
-	}
-
-	tk.MustExec("CREATE USER 'u6'@'localhost' IDENTIFIED BY 'password' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 3;")
-	tk.MustQuery(" SHOW CREATE USER 'u6'@'localhost';").Check(
-		testkit.Rows("CREATE USER 'u6'@'localhost' IDENTIFIED WITH 'mysql_native_password' AS '*2470C0C06DEE42FD1618BB99005ADCA2EC9D1E19' REQUIRE NONE PASSWORD EXPIRE DEFAULT ACCOUNT UNLOCK PASSWORD HISTORY DEFAULT PASSWORD REUSE INTERVAL DEFAULT FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 3"))
-
-	tk.MustExec("CREATE USER 'u7'@'localhost' IDENTIFIED BY 'password';")
-	tk.MustQuery(" SHOW CREATE USER 'u7'@'localhost';").Check(
-		testkit.Rows("CREATE USER 'u7'@'localhost' IDENTIFIED WITH 'mysql_native_password' AS '*2470C0C06DEE42FD1618BB99005ADCA2EC9D1E19' REQUIRE NONE PASSWORD EXPIRE DEFAULT ACCOUNT UNLOCK PASSWORD HISTORY DEFAULT PASSWORD REUSE INTERVAL DEFAULT"))
-
-	tk.MustExec("CREATE USER 'u8'@'localhost' IDENTIFIED BY 'password' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME UNBOUNDED;")
-	tk.MustQuery(" SHOW CREATE USER 'u8'@'localhost';").Check(
-		testkit.Rows("CREATE USER 'u8'@'localhost' IDENTIFIED WITH 'mysql_native_password' AS '*2470C0C06DEE42FD1618BB99005ADCA2EC9D1E19' REQUIRE NONE PASSWORD EXPIRE DEFAULT ACCOUNT UNLOCK PASSWORD HISTORY DEFAULT PASSWORD REUSE INTERVAL DEFAULT FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME UNBOUNDED"))
-
-	tk.MustExec("ALTER USER 'u4'@'localhost' PASSWORD_LOCK_TIME 0 FAILED_LOGIN_ATTEMPTS 0")
-	tk.MustQuery("select user_attributes from mysql.user where user = 'u4' and host = 'localhost'").Check(testkit.Rows(`{"resource_group": "default"}`))
-	tk.MustExec("ALTER USER 'u4'@'localhost' account unlock")
-	tk.MustQuery("select user_attributes from mysql.user where user = 'u4' and host = 'localhost'").Check(testkit.Rows(`{"resource_group": "default"}`))
-	tk.MustExec("ALTER USER 'u4'@'localhost' PASSWORD_LOCK_TIME 6")
-	tk.MustQuery("select user_attributes from mysql.user where user = 'u4' and host = 'localhost'").Check(testkit.Rows(`{"Password_locking": {"failed_login_attempts": 0, "password_lock_time_days": 6}, "resource_group": "default"}`))
-}
-
 func TestSetPwd(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 	tk := testkit.NewTestKit(t, store)
@@ -1215,465 +1104,4 @@ func TestDropStatsForMultipleTable(t *testing.T) {
 	statsTbl2 = h.GetTableStats(tableInfo2)
 	require.True(t, statsTbl2.Pseudo)
 	h.SetLease(0)
-}
-
-func TestUserReuseControl(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	rootTK.MustQuery(`show variables like  "password_history"`).Check(testkit.Rows("password_history 0"))
-	rootTK.MustQuery(`show variables like  "password_reuse_interval"`).Check(testkit.Rows("password_reuse_interval 0"))
-	rootTK.MustExec(`set global password_history = -1`)
-	rootTK.MustExec(`set global password_reuse_interval = -1`)
-	rootTK.MustQuery(`show variables like  "password_history"`).Check(testkit.Rows("password_history 0"))
-	rootTK.MustQuery(`show variables like  "password_reuse_interval"`).Check(testkit.Rows("password_reuse_interval 0"))
-	rootTK.MustExec(`set global password_history = 4294967295`)
-	rootTK.MustExec(`set global password_reuse_interval = 4294967295`)
-	rootTK.MustQuery(`show variables like  "password_history"`).Check(testkit.Rows("password_history 4294967295"))
-	rootTK.MustQuery(`show variables like  "password_reuse_interval"`).Check(testkit.Rows("password_reuse_interval 4294967295"))
-	rootTK.MustExec(`set global password_history = 4294967296`)
-	rootTK.MustExec(`set global password_reuse_interval = 4294967296`)
-	rootTK.MustQuery(`show variables like  "password_history"`).Check(testkit.Rows("password_history 4294967295"))
-	rootTK.MustQuery(`show variables like  "password_reuse_interval"`).Check(testkit.Rows("password_reuse_interval 4294967295"))
-	rootTK.MustGetErrCode(`set session password_history = 42949`, 1229)
-	rootTK.MustGetErrCode(`set session password_reuse_interval = 42949`, 1229)
-}
-
-func TestUserReuseInfo(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	rootTK.MustExec(`CREATE USER testReuse`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`<nil> <nil>`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD HISTORY 5`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`5 <nil>`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD HISTORY 0`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`0 <nil>`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD HISTORY DEFAULT`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`<nil> <nil>`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD HISTORY 65536`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`65535 <nil>`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD REUSE INTERVAL 5 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`65535 5`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD REUSE INTERVAL 0 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`65535 0`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD REUSE INTERVAL DEFAULT`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`65535 <nil>`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD REUSE INTERVAL 65536 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`65535 65535`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD HISTORY 6 PASSWORD REUSE INTERVAL 6 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`6 6`))
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD HISTORY 6 PASSWORD HISTORY 7 `)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`7 6`))
-
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD HISTORY 5`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`5 <nil>`))
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD REUSE INTERVAL 5 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`<nil> 5`))
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD REUSE INTERVAL 5 DAY PASSWORD REUSE INTERVAL 6 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`<nil> 6`))
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD HISTORY 5 PASSWORD REUSE INTERVAL 6 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`5 6`))
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD REUSE INTERVAL 6 DAY PASSWORD HISTORY 5`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`5 6`))
-
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustGetErrCode(`CREATE USER testReuse PASSWORD HISTORY -5`, 1064)
-	rootTK.MustGetErrCode(`CREATE USER testReuse PASSWORD REUSE INTERVAL -6 DAY`, 1064)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD HISTORY 65535 PASSWORD REUSE INTERVAL 65535 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`65535 65535`))
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD HISTORY 65536 PASSWORD REUSE INTERVAL 65536 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`65535 65535`))
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD HISTORY DEFAULT PASSWORD REUSE INTERVAL DEFAULT`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`<nil> <nil>`))
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD HISTORY 0 PASSWORD REUSE INTERVAL 0 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(`0 0`))
-}
-
-func TestUserReuseFunction(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustExec(`set global password_history = 1;`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustExec(`DROP USER testReuse`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-
-	rootTK.MustExec(`set global password_history = 0;`)
-	rootTK.MustExec(`set global password_reuse_interval = 1;`)
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`2`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test2'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`3`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test3'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`4`))
-	rootTK.MustExec(`update mysql.password_history set Password_timestamp = date_sub(Password_timestamp,interval '1 0:0:1' DAY_SECOND)`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustExec(`drop USER testReuse `)
-
-	rootTK.MustExec(`set global password_reuse_interval = 0;`)
-	//password nil is not stored
-	rootTK.MustExec(`CREATE USER testReuse PASSWORD HISTORY 5 PASSWORD REUSE INTERVAL 6 DAY`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustExec(`drop USER testReuse `)
-
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test' PASSWORD HISTORY 5`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test2'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test3'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`4`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustExec(`alter USER testReuse identified by 'test4'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`5`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test5'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`5`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test1'`, 3638)
-	rootTK.MustExec(`alter USER testReuse identified by 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`5`))
-	rootTK.MustExec(`drop USER testReuse`)
-
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test' PASSWORD HISTORY 5 PASSWORD REUSE INTERVAL 3 DAY`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test2'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test3'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test4'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test5'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`6`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustExec(`update mysql.password_history set Password_timestamp = date_sub(Password_timestamp,interval '3 0:0:1' DAY_SECOND) where user = 'testReuse' order by Password_timestamp asc limit 1`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`6`))
-	rootTK.MustExec(`drop USER testReuse`)
-
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test' PASSWORD HISTORY 5 PASSWORD REUSE INTERVAL 3 DAY`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test2'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test3'`)
-	rootTK.MustExec(`update mysql.password_history set Password_timestamp = date_sub(Password_timestamp,interval '3 0:0:1' DAY_SECOND) where user = 'testReuse' order by Password_timestamp asc limit 1`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`4`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustExec(`ALTER USER testReuse PASSWORD HISTORY 3`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`4`))
-	rootTK.MustExec(`drop USER testReuse`)
-
-	rootTK.MustExec(`set global password_history = 1;`)
-	rootTK.MustExec(`set global password_reuse_interval = 1;`)
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test' PASSWORD HISTORY 0 PASSWORD REUSE INTERVAL 0 DAY`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustExec(`drop USER testReuse`)
-
-	rootTK.MustExec(`set global password_history = 0;`)
-	rootTK.MustExec(`set global password_reuse_interval = 360000000;`)
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustGetErrCode(`set PASSWORD FOR testReuse = 'test'`, 3638)
-	rootTK.MustExec(`alter USER testReuse identified by ''`)
-	rootTK.MustExec(`alter USER testReuse identified by ''`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`2`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test2'`)
-	rootTK.MustExec(`set global password_reuse_interval = 4294967295;`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test3'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`4`))
-	rootTK.MustExec(`set PASSWORD FOR testReuse = 'test4'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`5`))
-	rootTK.MustExec(`drop USER testReuse`)
-
-	rootTK.MustExec(`set global password_reuse_interval = 0;`)
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test' PASSWORD HISTORY 5`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`2`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test1' PASSWORD HISTORY 0`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test1' PASSWORD HISTORY 2 PASSWORD REUSE INTERVAL 1 DAY`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test2'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test3'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test1' PASSWORD HISTORY 2 PASSWORD REUSE INTERVAL 0 DAY`)
-
-	// Support password and default value modification at the same time.
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`set global password_history = 1`)
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test' PASSWORD HISTORY DEFAULT PASSWORD REUSE INTERVAL DEFAULT`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustGetErrCode(`ALTER USER testReuse identified by 'test' PASSWORD HISTORY DEFAULT PASSWORD REUSE INTERVAL DEFAULT`, 3638)
-	rootTK.MustExec(`ALTER USER testReuse identified by 'test1' PASSWORD HISTORY DEFAULT PASSWORD REUSE INTERVAL DEFAULT`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-}
-
-func TestUserReuseDifferentAuth(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	// test caching_sha2_password.
-	rootTK.MustExec(`CREATE USER testReuse identified with 'caching_sha2_password' by 'test' PASSWORD HISTORY 1 `)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustGetErrCode(`set password for testReuse = 'test'`, 3638)
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustExec(`alter USER testReuse identified with 'tidb_sm3_password'`)
-	// changing the auth method prunes history.
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse identified with 'tidb_sm3_password' by 'test' PASSWORD HISTORY 1 `)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustGetErrCode(`set password for testReuse = 'test'`, 3638)
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustExec(`alter USER testReuse identified with 'caching_sha2_password'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustExec(`CREATE USER testReuse identified with 'caching_sha2_password' by 'test' PASSWORD REUSE INTERVAL 1 DAY`)
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustGetErrCode(`set password for testReuse = 'test'`, 3638)
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test2'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test3'`)
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustExec(`update mysql.password_history set Password_timestamp = date_sub(Password_timestamp,interval '1 0:0:1' DAY_SECOND) where user = 'testReuse' order by Password_timestamp asc limit 1`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test'`)
-
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustGetErrCode(`CREATE USER testReuse identified with 'mysql_clear_password' by 'test' PASSWORD REUSE INTERVAL 1 DAY`, 1524)
-	rootTK.MustGetErrCode(`CREATE USER testReuse identified with 'tidb_session_token' by 'test' PASSWORD REUSE INTERVAL 1 DAY`, 1524)
-	// no password.
-	rootTK.MustExec(`CREATE USER testReuse identified with 'auth_socket' by 'test' PASSWORD REUSE INTERVAL 1 DAY`)
-	rootTK.MustExec(`ALTER USER testReuse identified by 'test' `)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustQuery(`SELECT authentication_string FROM mysql.user WHERE user = 'testReuse'`).Check(testkit.Rows(""))
-	rootTK.MustExec(`ALTER USER testReuse identified with 'caching_sha2_password' by 'test' `)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	// AuthTiDBAuthToken is the token login method on the cloud,
-	// and the Password Reuse Policy does not take effect.
-	rootTK.MustExec(`drop USER testReuse`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustExec(`CREATE USER testReuse identified with 'tidb_auth_token' by 'test' PASSWORD REUSE INTERVAL 1 DAY`)
-	rootTK.MustExec(`ALTER USER testReuse identified by 'test' `)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustExec(`set password for testReuse = 'test'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustExec(`ALTER USER testReuse identified with 'caching_sha2_password' by 'test' `)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustGetErrCode(`alter USER testReuse identified by 'test'`, 3638)
-	rootTK.MustGetErrCode(`set password for testReuse = 'test'`, 3638)
-	rootTK.MustExec(`drop USER testReuse`)
-}
-
-func TestUserReuseMultiuser(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	//alter multi user success
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test', testReuse1 identified by 'test', testReuse2 identified by 'test' PASSWORD HISTORY 65535 PASSWORD REUSE INTERVAL 65535 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user like 'testReuse%'`).Check(testkit.Rows(`65535 65535`, `65535 65535`, `65535 65535`))
-	rootTK.MustExec(`ALTER USER testReuse identified by 'test1', testReuse1 identified by 'test1', testReuse2 identified by 'test1' PASSWORD HISTORY 3 PASSWORD REUSE INTERVAL 3 DAY`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user like 'testReuse%'`).Check(testkit.Rows(`3 3`, `3 3`, `3 3`))
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user like 'testReuse%' group by user`).Check(testkit.Rows(`2`, `2`, `2`))
-	//alter multi user fail
-	rootTK.MustExec(`CREATE USER testReuse3 identified by 'test'`)
-	rootTK.MustQuery(`SELECT Password_reuse_history,Password_reuse_time FROM mysql.user WHERE user like 'testReuse%'`).Check(testkit.Rows(`3 3`, `3 3`, `3 3`, `<nil> <nil>`))
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user like 'testReuse%' group by user`).Check(testkit.Rows(`2`, `2`, `2`))
-	rootTK.MustGetErrCode(`ALTER USER testReuse identified by 'test1', testReuse3 identified by 'test1'`, 3638)
-	//drop user
-	rootTK.MustExec(`drop User testReuse, testReuse1, testReuse2, testReuse3`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user like 'testReuse%' `).Check(testkit.Rows(`0`))
-}
-
-func TestUserReuseRename(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	rootTK.MustExec(`CREATE USER testReuse identified by 'test' PASSWORD HISTORY 5`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`1`))
-	rootTK.MustExec(`alter USER testReuse identified by 'test1'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test2'`)
-	rootTK.MustExec(`alter USER testReuse identified by 'test3'`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`4`))
-	rootTK.MustExec(`rename USER testReuse to testReuse1`)
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse'`).Check(testkit.Rows(`0`))
-	rootTK.MustQuery(`SELECT count(*) FROM mysql.password_history WHERE user = 'testReuse1'`).Check(testkit.Rows(`4`))
-}
-
-func TestUserAlterUser(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	rootTK.MustExec(`CREATE USER test1 IDENTIFIED WITH 'mysql_native_password' BY '1234'`)
-	alterUserSQL := `ALTER USER 'test1' IDENTIFIED BY '222', 'test_not_exist'@'localhost' IDENTIFIED BY '111';`
-	rootTK.MustGetErrCode(alterUserSQL, mysql.ErrCannotUser)
-	result := rootTK.MustQuery(`SELECT authentication_string FROM mysql.User WHERE User="test1" and Host="%"`)
-	result.Check(testkit.Rows(auth.EncodePassword("1234")))
-	alterUserSQL = `ALTER USER IF EXISTS 'test1' IDENTIFIED BY '222', 'test_not_exist'@'localhost' IDENTIFIED BY '111';`
-	rootTK.MustExec(alterUserSQL)
-	rootTK.MustQuery("show warnings").Check(testkit.RowsWithSep("|", "Note|3162|User 'test_not_exist'@'localhost' does not exist."))
-	result = rootTK.MustQuery(`SELECT authentication_string FROM mysql.User WHERE User="test1" and Host="%"`)
-	result.Check(testkit.Rows(auth.EncodePassword("222")))
-}
-
-func sha1Password(s string) []byte {
-	crypt := sha1.New()
-	crypt.Write([]byte(s))
-	hashStage1 := crypt.Sum(nil)
-	crypt.Reset()
-	crypt.Write(hashStage1)
-	hashStage2 := crypt.Sum(nil)
-	crypt.Reset()
-	crypt.Write(hashStage2)
-	hashStage3 := crypt.Sum(nil)
-	for i := range hashStage3 {
-		hashStage3[i] ^= hashStage1[i]
-	}
-	return hashStage3
-}
-
-func TestFailedLoginTrackingAlterUser(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	sql := new(strings.Builder)
-	checkUserAttributes := "select JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_attempts')," +
-		"JSON_EXTRACT(user_attributes, '$.Password_locking.auto_account_locked')," +
-		"JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_count')," +
-		"JSON_EXTRACT(user_attributes, '$.Password_locking.password_lock_time_days')," +
-		"JSON_EXTRACT(user_attributes, '$.metadata')from mysql.user where user= %? and host = %?"
-	tk := testkit.NewTestKit(t, store)
-	err := domain.GetDomain(rootTK.Session()).NotifyUpdatePrivilege()
-	require.NoError(t, err)
-	rootTK.MustExec(`CREATE USER test1 IDENTIFIED BY '1234' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 3 COMMENT 'test'`)
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "test1", Hostname: "%"}, sha1Password("1234"), nil)
-	require.NoError(t, err)
-	sqlexec.MustFormatSQL(sql, checkUserAttributes, "test1", "%")
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`3 <nil> <nil> 3 {"comment": "test"}`))
-	tk = testkit.NewTestKit(t, store)
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "test1", Hostname: "%"}, sha1Password("<wrong-password>"), nil)
-	require.Error(t, err)
-
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`3 "N" 1 3 {"comment": "test"}`))
-	rootTK.MustExec(`Alter user test1  FAILED_LOGIN_ATTEMPTS 4 `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 "N" 1 3 {"comment": "test"}`))
-	rootTK.MustExec(`Alter user test1  PASSWORD_LOCK_TIME 4 `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 "N" 1 4 {"comment": "test"}`))
-	rootTK.MustExec(`Alter user test1  COMMENT 'test1' `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 "N" 1 4 {"comment": "test1"}`))
-	rootTK.MustExec(`Alter user test1 FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 3 COMMENT 'test'`)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`3 "N" 1 3 {"comment": "test"}`))
-
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "test1", Hostname: "%"}, sha1Password("<wrong-password>"), nil)
-	require.Error(t, err)
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "test1", Hostname: "%"}, sha1Password("<wrong-password>"), nil)
-	require.Error(t, err)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`3 "Y" 3 3 {"comment": "test"}`))
-	rootTK.MustExec(`Alter user test1  FAILED_LOGIN_ATTEMPTS 4 `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 "Y" 3 3 {"comment": "test"}`))
-	rootTK.MustExec(`Alter user test1  PASSWORD_LOCK_TIME 4 `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 "Y" 3 4 {"comment": "test"}`))
-	rootTK.MustExec(`Alter user test1  COMMENT 'test2' `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 "Y" 3 4 {"comment": "test2"}`))
-	rootTK.MustExec(`Alter user test1  account unlock `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 "N" 0 4 {"comment": "test2"}`))
-
-	rootTK.MustExec(`Alter user test1  FAILED_LOGIN_ATTEMPTS 0 `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`0 "N" 0 4 {"comment": "test2"}`))
-	rootTK.MustExec(`Alter user test1  PASSWORD_LOCK_TIME 0 `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`<nil> <nil> <nil> <nil> {"comment": "test2"}`))
-
-	rootTK.MustExec(`Alter user test1  account unlock `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`<nil> <nil> <nil> <nil> {"comment": "test2"}`))
-	rootTK.MustExec(`Alter user test1  FAILED_LOGIN_ATTEMPTS 4 `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 <nil> <nil> 0 {"comment": "test2"}`))
-	rootTK.MustExec(`Alter user test1  account unlock `)
-	rootTK.MustQuery(sql.String()).Check(testkit.Rows(`4 "N" 0 0 {"comment": "test2"}`))
-}
-
-func TestMixPasswordPolicy(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	rootTK := testkit.NewTestKit(t, store)
-	tk := testkit.NewTestKit(t, store)
-	// Password Strength Check.
-	rootTK.MustExec(`set global validate_password.enable = ON`)
-	rootTK.MustGetErrCode(`create user u2 identified by 'u2' PASSWORD EXPIRE INTERVAL 2 DAY password history 2
-		password reuse interval 2 day FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 1`, 1819)
-	rootTK.MustGetErrCode(`create user u2`, 1819)
-	rootTK.MustGetErrCode(`create user u2 identified by 'u2222222' PASSWORD EXPIRE INTERVAL 2 DAY password history 2
-		password reuse interval 2 day FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 1`, 1819)
-	rootTK.MustGetErrCode(`create user u2 identified by 'Uu2222222' PASSWORD EXPIRE INTERVAL 2 DAY password history 2
-		password reuse interval 2 day FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 1`, 1819)
-	rootTK.MustGetErrCode(`create user u2 identified by 'Uu3222222' PASSWORD EXPIRE INTERVAL 2 DAY password history 2
-		password reuse interval 2 day FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 1`, 1819)
-	rootTK.MustExec(`create user u2 identified by 'Uu3@22222' PASSWORD EXPIRE INTERVAL 2 DAY password history 2
-		password reuse interval 2 day FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 1`)
-	rootTK.MustQuery(`Select count(*) from mysql.password_history where user = 'u2' and host = '%'`).Check(testkit.Rows("1"))
-	result := rootTK.MustQuery(`Select authentication_string from mysql.user where user = 'u2' and host = '%'`)
-	result.Check(testkit.Rows(auth.EncodePassword("Uu3@22222")))
-	// Disable password reuse.
-	rootTK.MustGetErrCode(`Alter user u2 identified by 'Uu3@22222'`, 3638)
-	rootTK.MustGetErrCode(`Set password for 'u2' = 'Uu3@22222'`, 3638)
-	// Password Strength Check.
-	rootTK.MustGetErrCode(`Alter user u2 identified by 'U2'`, 1819)
-	rootTK.MustGetErrCode(`Set password for 'u2' = 'U2'`, 1819)
-	// Did not modify successfully.
-	result = rootTK.MustQuery(`Select authentication_string from mysql.user where user = 'u2' and host = '%'`)
-	result.Check(testkit.Rows(auth.EncodePassword("Uu3@22222")))
-	// Auto-lock in effect.
-	err := tk.Session().Auth(&auth.UserIdentity{Username: "u2", Hostname: "%"}, sha1Password("<wrong-password>"), nil)
-	require.ErrorContains(t, err, "Account is blocked for 1 day(s) (1 day(s) remaining) due to 1 consecutive failed logins.")
-	result = rootTK.MustQuery(`SELECT
-	JSON_UNQUOTE(JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_count')),
-	JSON_UNQUOTE(JSON_EXTRACT(user_attributes, '$.Password_locking.auto_account_locked')) from mysql.user where user = 'u2' and host = '%'`)
-	result.Check(testkit.Rows(`1 Y`))
-	rootTK.MustExec(`ALTER user u2 account unlock`)
-
-	// Unlock in effect.
-	result = rootTK.MustQuery(`SELECT
-	JSON_UNQUOTE(JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_count')),
-	JSON_UNQUOTE(JSON_EXTRACT(user_attributes, '$.Password_locking.auto_account_locked')) from mysql.user where user = 'u2' and host = '%'`)
-	result.Check(testkit.Rows(`0 N`))
-
-	rootTK.MustExec(`set global validate_password.enable = OFF`)
-	rootTK.MustExec(`update mysql.user set Password_last_changed = date_sub(Password_last_changed,interval '3 0:0:1' DAY_SECOND)  where user = 'u2' and host = '%'`)
-	err = domain.GetDomain(rootTK.Session()).NotifyUpdatePrivilege()
-	require.NoError(t, err)
-	// Password expires and takes effect.
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "u2", Hostname: "%"}, sha1Password("Uu3@22222"), nil)
-	require.ErrorContains(t, err, "Your password has expired.")
-	variable.IsSandBoxModeEnabled.Store(true)
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "u2", Hostname: "%"}, sha1Password("Uu3@22222"), nil)
-	require.NoError(t, err)
-	require.True(t, tk.Session().InSandBoxMode())
-
-	rootTK.MustExec(`set global validate_password.enable = ON`)
-	// Forbid other users to change password.
-	tk.MustGetErrCode(`Alter user root identified by 'Uu3@22222'`, 1820)
-	// Disable password reuse.
-	tk.MustGetErrCode(`Alter user u2 identified by 'Uu3@22222'`, 3638)
-	tk.MustGetErrCode(`set password = 'Uu3@22222'`, 3638)
-	// Password Strength Check.
-	tk.MustGetErrCode(`Alter user u2 identified by 'U2'`, 1819)
-	tk.MustGetErrCode(`set password = 'U2'`, 1819)
-	tk.MustExec(`Set password = 'Uu3@22223'`)
-	require.False(t, tk.Session().InSandBoxMode())
-	rootTK.MustQuery(`Select count(*) from mysql.password_history where user = 'u2' and host = '%'`).Check(testkit.Rows("2"))
-	result = rootTK.MustQuery(`Select authentication_string from mysql.user where user = 'u2' and host = '%'`)
-	result.Check(testkit.Rows(auth.EncodePassword("Uu3@22223")))
-	tk = testkit.NewTestKit(t, store)
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "u2", Hostname: "%"}, sha1Password("Uu3@22223"), nil)
-	require.NoError(t, err)
 }
