+++ privilege/privileges/privileges_test.go
@@ -17,21 +17,17 @@ package privileges_test
 import (
 	"bytes"
 	"context"
-	"crypto/sha1"
 	"crypto/tls"
 	"crypto/x509"
 	"crypto/x509/pkix"
-	"encoding/json"
 	"fmt"
 	"net/url"
 	"os"
-	"strconv"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/pingcap/tidb/config"
-	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/executor"
 	"github.com/pingcap/tidb/kv"
@@ -3107,579 +3103,4 @@ func TestPasswordExpireWithSandBoxMode(t *testing.T) {
 	err = tk.Session().Auth(user, nil, nil)
 	require.NoError(t, err)
 	require.False(t, tk.Session().InSandBoxMode())
-}
-
-func TestFailedLoginTrackingCheckError(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-
-	// Set FAILED_LOGIN_ATTEMPTS to 1, and check error messages after  login failure once.
-	createAndCheck(tk, "CREATE USER 'testu1'@'localhost' IDENTIFIED BY 'testu1' FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 1",
-		"{\"Password_locking\": {\"failed_login_attempts\": 1, \"password_lock_time_days\": 1}, \"resource_group\": \"default\"}", "testu1")
-	err := tk.Session().Auth(&auth.UserIdentity{Username: "testu1", Hostname: "localhost"}, sha1Password("password"), nil)
-	lds := strconv.FormatInt(1, 10)
-	errTarget := privileges.GenerateAccountAutoLockErr(1, "testu1", "localhost", lds, lds)
-	require.Equal(t, err.Error(), errTarget.Error())
-	checkAuthUser(t, tk, "testu1", 1, "Y")
-
-	// Check the login error message after the account is locked.
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "testu1", Hostname: "localhost"}, sha1Password("password"), nil)
-	require.Equal(t, err.Error(), errTarget.Error())
-	checkAuthUser(t, tk, "testu1", 1, "Y")
-
-	// Set FAILED_LOGIN_ATTEMPTS to 1 and PASSWORD_LOCK_TIME to UNBOUNDED. Check error messages after failed login once.
-	createAndCheck(tk, "CREATE USER 'testu2'@'localhost' IDENTIFIED BY 'testu2' FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME UNBOUNDED",
-		"{\"Password_locking\": {\"failed_login_attempts\": 1, \"password_lock_time_days\": -1}, \"resource_group\": \"default\"}", "testu2")
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "testu2", Hostname: "localhost"}, sha1Password("password"), nil)
-	errTarget = privileges.GenerateAccountAutoLockErr(1, "testu2", "localhost", "unlimited", "unlimited")
-	require.Equal(t, err.Error(), errTarget.Error())
-	checkAuthUser(t, tk, "testu2", 1, "Y")
-
-	// Check the login error message after the account is locked.
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "testu2", Hostname: "localhost"}, sha1Password("password"), nil)
-	require.Equal(t, err.Error(), errTarget.Error())
-	checkAuthUser(t, tk, "testu2", 1, "Y")
-
-	// Set FAILED_LOGIN_ATTEMPTS to 0 or PASSWORD_LOCK_TIME to 0. Check error messages after failed login once.
-	createAndCheck(tk, "CREATE USER 'testu3'@'localhost' IDENTIFIED BY 'testu3' FAILED_LOGIN_ATTEMPTS 0 PASSWORD_LOCK_TIME UNBOUNDED",
-		"{\"Password_locking\": {\"failed_login_attempts\": 0, \"password_lock_time_days\": -1}, \"resource_group\": \"default\"}", "testu3")
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "testu3", Hostname: "localhost"}, sha1Password("password"), nil)
-	require.True(t, strings.Contains(err.Error(), "Access denied for user 'testu3'@'localhost' (using password: YES)"))
-	checkAuthUser(t, tk, "testu3", 0, "")
-	createAndCheck(tk, "CREATE USER 'testu4'@'localhost' IDENTIFIED BY 'testu4' FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 0",
-		"{\"Password_locking\": {\"failed_login_attempts\": 1, \"password_lock_time_days\": 0}, \"resource_group\": \"default\"}", "testu4")
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "testu4", Hostname: "localhost"}, sha1Password("password"), nil)
-	require.True(t, strings.Contains(err.Error(), "Access denied for user 'testu4'@'localhost' (using password: YES)"))
-	checkAuthUser(t, tk, "testu4", 0, "")
-	tk.MustExec("CREATE USER 'testu5'@'localhost' IDENTIFIED BY 'testu5' FAILED_LOGIN_ATTEMPTS 0 PASSWORD_LOCK_TIME 0")
-	err = tk.Session().Auth(&auth.UserIdentity{Username: "testu5", Hostname: "localhost"}, sha1Password("password"), nil)
-	require.True(t, strings.Contains(err.Error(), "Access denied for user 'testu5'@'localhost' (using password: YES)"))
-	tk.MustQuery("select user_attributes from mysql.user where user= 'testu5' and host = 'localhost'").Check(testkit.Rows("{\"resource_group\": \"default\"}"))
-}
-
-func TestFailedLoginTrackingWithSpecifiedUserUserAttributes(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-
-	// Create user specifying only comment.
-	tk.MustExec("CREATE USER 'testu1'@'localhost' IDENTIFIED BY 'testu1' comment 'testcomment' ")
-	tk.MustQuery("select user_attributes from mysql.user where user= 'testu1' and host = 'localhost'").
-		Check(testkit.Rows("{\"metadata\": {\"comment\": \"testcomment\"}, \"resource_group\": \"default\"}"))
-
-	// Create user specifying only attribute.
-	tk.MustExec("create user testu2@'localhost' identified by 'testu2' ATTRIBUTE '{\"attribute\":\"testattribute\"}'")
-	tk.MustQuery("select user_attributes from mysql.user where user= 'testu2' and host = 'localhost'").
-		Check(testkit.Rows("{\"metadata\": {\"attribute\": \"testattribute\"}, \"resource_group\": \"default\"}"))
-
-	// Create user specified comment and FAILED_LOGIN_ATTEMPTS and PASSWORD_LOCK_TIME.
-	tk.MustExec("create user testu3@'localhost' identified by 'testu3' FAILED_LOGIN_ATTEMPTS 1 " +
-		"PASSWORD_LOCK_TIME 1 comment 'testcomment'")
-	checkUserUserAttributes(tk, "testu3", "localhost", "1 <nil> <nil> 1 {\"comment\": \"testcomment\"}")
-
-	// Create user specified attribute and FAILED_LOGIN_ATTEMPTS and PASSWORD_LOCK_TIME.
-	tk.MustExec("create user testu4@'localhost' identified by 'testu4' FAILED_LOGIN_ATTEMPTS 1 " +
-		"PASSWORD_LOCK_TIME 1 ATTRIBUTE '{\"attribute\":\"testattribute\"}'")
-	checkUserUserAttributes(tk, "testu4", "localhost", "1 <nil> <nil> 1 {\"attribute\": \"testattribute\"}")
-
-	// Create user specified comment, FAILED_LOGIN_ATTEMPTS, and PASSWORD_LOCK_TIME,
-	// and confirm the user_attributes column value after login fails.
-	tk.MustExec("create user testu5@'localhost' identified by 'testu5' FAILED_LOGIN_ATTEMPTS 2 " +
-		"PASSWORD_LOCK_TIME 1 comment 'testcomment'")
-	checkUserUserAttributes(tk, "testu5", "localhost", "2 <nil> <nil> 1 {\"comment\": \"testcomment\"}")
-
-	// Confirm the user_attributes value after login failure once.
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu5", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu5", "localhost", "2 \"N\" 1 1 {\"comment\": \"testcomment\"}")
-
-	// After the number of failed login attempts reaches FAILED_LOGIN_ATTEMPTS, check the account lock status.
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu5", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu5", "localhost", "2 \"Y\" 2 1 {\"comment\": \"testcomment\"}")
-	// After the account is locked, manually unlock the account and check the user_attributes value.
-	tk.MustExec("alter user testu5@'localhost' account  unlock")
-	checkUserUserAttributes(tk, "testu5", "localhost", "2 \"N\" 0 1 {\"comment\": \"testcomment\"}")
-
-	// Create user specified comment, FAILED_LOGIN_ATTEMPTS, and PASSWORD_LOCK_TIME,
-	// and confirm the user_attributes column value after login fails.
-	tk.MustExec("create user testu6@'localhost' identified by '' FAILED_LOGIN_ATTEMPTS 2 PASSWORD_LOCK_TIME 1 " +
-		"comment 'testcomment'")
-	// Confirm the user_attributes value after login failure once.
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu6", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu6", "localhost", "2 \"N\" 1 1 {\"comment\": \"testcomment\"}")
-
-	// After the number of failed login attempts reaches FAILED_LOGIN_ATTEMPTS, check the account lock status.
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu6", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu6", "localhost", "2 \"Y\" 2 1 {\"comment\": \"testcomment\"}")
-
-	// After the account is automatically locked, change the lock time and check
-	// the user_attributes value after logging in successfully.
-	changeAutoLockedLastChanged(tk, "-72h1s", "testu6")
-	sk1 := testkit.NewTestKit(t, store)
-	require.NoError(t, sk1.Session().Auth(&auth.UserIdentity{Username: "testu6", Hostname: "localhost"}, nil, nil))
-	checkUserUserAttributes(tk, "testu6", "localhost", "3 \"N\" 0 3 {\"comment\": \"testcomment\"}")
-
-	// Create user specified attributes, FAILED_LOGIN_ATTEMPTS, and PASSWORD_LOCK_TIME,
-	// and confirm the user_attributes column value after login fails.
-	tk.MustExec("create user testu7@'localhost' identified by 'testu7' FAILED_LOGIN_ATTEMPTS 2 PASSWORD_LOCK_TIME 1 " +
-		"ATTRIBUTE '{\"attribute\":\"testattribute\"}'")
-	checkUserUserAttributes(tk, "testu7", "localhost", "2 <nil> <nil> 1 {\"attribute\": \"testattribute\"}")
-
-	// Confirm the user_attributes value after login failure once.
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu7", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu7", "localhost", "2 \"N\" 1 1 {\"attribute\": \"testattribute\"}")
-
-	// After the number of failed login attempts reaches FAILED_LOGIN_ATTEMPTS, check the account lock status.
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu7", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu7", "localhost", "2 \"Y\" 2 1 {\"attribute\": \"testattribute\"}")
-
-	// After the account is locked, manually unlock the account and check the user_attributes value.
-	tk.MustExec("alter user testu7@'localhost' account  unlock")
-	checkUserUserAttributes(tk, "testu7", "localhost", "2 \"N\" 0 1 {\"attribute\": \"testattribute\"}")
-
-	tk.MustExec("create user testu8@'localhost' identified by '' FAILED_LOGIN_ATTEMPTS 2 PASSWORD_LOCK_TIME 1" +
-		" ATTRIBUTE '{\"attribute\":\"testattribute\"}'")
-	// Confirm the user_attributes value after login failure once.
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu8", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu8", "localhost", "2 \"N\" 1 1 {\"attribute\": \"testattribute\"}")
-
-	// After the number of failed login attempts reaches FAILED_LOGIN_ATTEMPTS, check the account lock status.
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu8", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu8", "localhost", "2 \"Y\" 2 1 {\"attribute\": \"testattribute\"}")
-
-	// After the account is automatically locked, change the lock time and check
-	// the user_attributes value after logging in successfully.
-	changeAutoLockedLastChanged(tk, "-72h1s", "testu8")
-	sk2 := testkit.NewTestKit(t, store)
-	require.NoError(t, sk2.Session().Auth(&auth.UserIdentity{Username: "testu8", Hostname: "localhost"}, nil, nil))
-	checkUserUserAttributes(tk, "testu8", "localhost", "3 \"N\" 0 3 {\"attribute\": \"testattribute\"}")
-
-	// FAILED_LOGIN_ATTEMPTS is set to 2 . check user_attributes value after
-	// the user login fails once ,and login success at second time.
-	tk.MustExec("create user testu9@'localhost' identified by '' FAILED_LOGIN_ATTEMPTS 2 PASSWORD_LOCK_TIME 1" +
-		" comment 'testcomment'")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu9", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu9", "localhost", "2 \"N\" 1 1 {\"comment\": \"testcomment\"}")
-	sk3 := testkit.NewTestKit(t, store)
-	require.NoError(t, sk3.Session().Auth(&auth.UserIdentity{Username: "testu9", Hostname: "localhost"},
-		nil, nil))
-	checkUserUserAttributes(tk, "testu9", "localhost", "2 \"N\" 0 1 {\"comment\": \"testcomment\"}")
-
-	// FAILED_LOGIN_ATTEMPTS or PASSWORD_LOCK_TIME is set to 0. Check user_attributes value after login fail.
-	tk.MustExec("create user testu10@'localhost' identified by '' FAILED_LOGIN_ATTEMPTS 2 PASSWORD_LOCK_TIME 0 " +
-		"comment 'testcomment'")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu10", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu10", "localhost", "2 <nil> <nil> 0 {\"comment\": \"testcomment\"}")
-
-	tk.MustExec("create user testu11@'localhost' identified by '' FAILED_LOGIN_ATTEMPTS 0 PASSWORD_LOCK_TIME 2 " +
-		"comment 'testcomment'")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu11", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu11", "localhost", "0 <nil> <nil> 2 {\"comment\": \"testcomment\"}")
-
-	// The account is automatically locked after the user specifies FAILED_LOGIN_ATTEMPTS and PASSWORD_LOCK_TIME.
-	// Change FAILED_LOGIN_ATTEMPTS or PASSWORD_LOCK_TIME to 0, and check whether the user can login.
-	tk.MustExec("create user testu12@'localhost' identified by '' FAILED_LOGIN_ATTEMPTS 2 PASSWORD_LOCK_TIME 1 " +
-		"comment 'testcomment'")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu12", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu12", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu12", "localhost", "2 \"Y\" 2 1 {\"comment\": \"testcomment\"}")
-	tk.MustExec("alter user testu12@'localhost' FAILED_LOGIN_ATTEMPTS 0")
-	checkUserUserAttributes(tk, "testu12", "localhost", "0 \"Y\" 2 1 {\"comment\": \"testcomment\"}")
-	sk4 := testkit.NewTestKit(t, store)
-	require.NoError(t, sk4.Session().Auth(&auth.UserIdentity{Username: "testu12", Hostname: "localhost"},
-		nil, nil))
-}
-
-func TestFailedLoginTrackingAlterUser(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-
-	// Create user specifying only comment.
-	tk.MustExec("CREATE USER 'testu1'@'localhost' IDENTIFIED BY 'testu1' comment 'testcomment' ")
-	tk.MustQuery("select user_attributes from mysql.user where user= 'testu1' and host = 'localhost'").
-		Check(testkit.Rows("{\"metadata\": {\"comment\": \"testcomment\"}, \"resource_group\": \"default\"}"))
-	tk.MustExec("Alter USER 'testu1'@'localhost' comment ''")
-	tk.MustQuery("select user_attributes from mysql.user where user= 'testu1' and host = 'localhost'").
-		Check(testkit.Rows("{\"metadata\": {\"comment\": \"\"}, \"resource_group\": \"default\"}"))
-
-	// Create user specifying only attribute.
-	tk.MustExec("CREATE USER 'testu2'@'localhost' IDENTIFIED BY 'testu2' ATTRIBUTE '{\"attribute\":\"testattribute\"}'")
-	tk.MustQuery("select user_attributes from mysql.user where user= 'testu2' and host = 'localhost'").
-		Check(testkit.Rows("{\"metadata\": {\"attribute\": \"testattribute\"}, \"resource_group\": \"default\"}"))
-	tk.MustExec("Alter USER 'testu2'@'localhost' ATTRIBUTE '{\"attribute\":\"test\"}'")
-	tk.MustQuery("select user_attributes from mysql.user where user= 'testu2' and host = 'localhost'").
-		Check(testkit.Rows("{\"metadata\": {\"attribute\": \"test\"}, \"resource_group\": \"default\"}"))
-
-	// Create a user and specify FAILED_LOGIN_ATTEMPTS, PASSWORD_LOCK_TIME, and COMMENT.
-	// Check the user_attributes value after alter user.
-	tk.MustExec("CREATE USER 'testu3'@'localhost' IDENTIFIED BY 'testu3' FAILED_LOGIN_ATTEMPTS 1 " +
-		"PASSWORD_LOCK_TIME 1 comment 'testcomment'")
-	checkUserUserAttributes(tk, "testu3", "localhost", "1 <nil> <nil> 1 {\"comment\": \"testcomment\"}")
-	tk.MustExec("alter user 'testu3'@'localhost' FAILED_LOGIN_ATTEMPTS 0")
-	checkUserUserAttributes(tk, "testu3", "localhost", "0 <nil> <nil> 1 {\"comment\": \"testcomment\"}")
-	tk.MustExec("alter user 'testu3'@'localhost' PASSWORD_LOCK_TIME 0")
-	tk.MustQuery("select JSON_EXTRACT(user_attributes, '$.Password_locking')," +
-		"JSON_EXTRACT(user_attributes, '$.metadata')from mysql.user where user= 'testu3' and host = 'localhost'").
-		Check(testkit.Rows("<nil> {\"comment\": \"testcomment\"}"))
-
-	// Create a user and specify FAILED_LOGIN_ATTEMPTS, PASSWORD_LOCK_TIME, and ATTRIBUTE.
-	// Check the user_attributes value after alter user.
-	tk.MustExec("CREATE USER 'testu4'@'localhost' IDENTIFIED BY 'testu4' FAILED_LOGIN_ATTEMPTS 1 " +
-		"PASSWORD_LOCK_TIME 1 ATTRIBUTE '{\"attribute\":\"testattribute\"}'")
-	checkUserUserAttributes(tk, "testu4", "localhost", "1 <nil> <nil> 1 {\"attribute\": \"testattribute\"}")
-	tk.MustExec("alter user 'testu4'@'localhost' FAILED_LOGIN_ATTEMPTS 0")
-	checkUserUserAttributes(tk, "testu4", "localhost", "0 <nil> <nil> 1 {\"attribute\": \"testattribute\"}")
-	tk.MustExec("alter user 'testu4'@'localhost' PASSWORD_LOCK_TIME 0")
-	tk.MustQuery("select JSON_EXTRACT(user_attributes, '$.Password_locking')," +
-		"JSON_EXTRACT(user_attributes, '$.metadata')from mysql.user where user= 'testu4' and host = 'localhost'").
-		Check(testkit.Rows("<nil> {\"attribute\": \"testattribute\"}"))
-
-	// Create a user and specify FAILED_LOGIN_ATTEMPTS, PASSWORD_LOCK_TIME, and ATTRIBUTE.
-	// Check the user_attributes value after alter user.
-	tk.MustExec("CREATE USER 'testu5'@'localhost' IDENTIFIED BY 'testu5' FAILED_LOGIN_ATTEMPTS 1 " +
-		"PASSWORD_LOCK_TIME 1 ATTRIBUTE '{\"attribute\":\"testattribute\"}'")
-	checkUserUserAttributes(tk, "testu5", "localhost", "1 <nil> <nil> 1 {\"attribute\": \"testattribute\"}")
-	tk.MustExec("alter user 'testu5'@'localhost' FAILED_LOGIN_ATTEMPTS 0 PASSWORD_LOCK_TIME 0  ATTRIBUTE '{\"attribute\":\"test\"}'")
-	tk.MustQuery("select JSON_EXTRACT(user_attributes, '$.Password_locking')," +
-		"JSON_EXTRACT(user_attributes, '$.metadata')from mysql.user where user= 'testu5' and host = 'localhost'").
-		Check(testkit.Rows("<nil> {\"attribute\": \"test\"}"))
-
-	// Create a user to specify a comment, modify the user to add an ATTRIBUTE,
-	// modify the user to delete a comment, and check the user_attributes value.
-	tk.MustExec("CREATE USER 'testu6'@'localhost' IDENTIFIED BY 'testu6' FAILED_LOGIN_ATTEMPTS 1 " +
-		"PASSWORD_LOCK_TIME 1 comment 'testcomment'")
-	checkUserUserAttributes(tk, "testu6", "localhost", "1 <nil> <nil> 1 {\"comment\": \"testcomment\"}")
-	tk.MustExec("alter user 'testu6'@'localhost' ATTRIBUTE '{\"attribute\": \"testattribute\"}'")
-	checkUserUserAttributes(tk, "testu6", "localhost", "1 <nil> <nil> 1 {\"attribute\": \"testattribute\", \"comment\": \"testcomment\"}")
-	tk.MustExec("alter user 'testu6'@'localhost' ATTRIBUTE '{\"comment\": null}'")
-	checkUserUserAttributes(tk, "testu6", "localhost", "1 <nil> <nil> 1 {\"attribute\": \"testattribute\"}")
-
-	// After consecutive login failures and the account is locked,
-	// change the values of FAILED_LOGIN_ATTEMPTS and PASSWORD_LOCK_TIME to 0 and check the user_attributes value
-	tk.MustExec("CREATE USER 'testu7'@'localhost' IDENTIFIED BY 'testu7' FAILED_LOGIN_ATTEMPTS 1 " +
-		"PASSWORD_LOCK_TIME 1 comment 'testcomment'")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu7", Hostname: "localhost"},
-		sha1Password("password"), nil))
-	checkUserUserAttributes(tk, "testu7", "localhost", "1 \"Y\" 1 1 {\"comment\": \"testcomment\"}")
-	tk.MustExec("alter user 'testu7'@'localhost' FAILED_LOGIN_ATTEMPTS 0 PASSWORD_LOCK_TIME 0")
-	tk.MustQuery("select JSON_EXTRACT(user_attributes, '$.Password_locking'),JSON_EXTRACT(user_attributes,'$.metadata') " +
-		"from mysql.user where user='testu7' and host ='localhost'").Check(testkit.Rows("<nil> {\"comment\": \"testcomment\"}"))
-
-	// Create a user and specify FAILED_LOGIN_ATTEMPTS, PASSWORD_LOCK_TIME.
-	// Check the user_attributes value after alter user.
-	tk.MustExec("CREATE USER 'testu8'@'localhost' IDENTIFIED BY 'testu5' FAILED_LOGIN_ATTEMPTS 1 " +
-		"PASSWORD_LOCK_TIME 1")
-	checkUserUserAttributes(tk, "testu8", "localhost", "1 <nil> <nil> 1 <nil>")
-	tk.MustExec("alter user 'testu8'@'localhost' FAILED_LOGIN_ATTEMPTS 0 PASSWORD_LOCK_TIME 0")
-	tk.MustQuery("select user_attributes from mysql.user where user= 'testu8' and host = 'localhost'").
-		Check(testkit.Rows("{\"resource_group\": \"default\"}"))
-
-	// Specify only FAILED_LOGIN_ATTEMPTS one attribute when creating user.
-	// Change the value to 0 and check the user_attributes value.
-	tk.MustExec("CREATE USER 'testu9'@'localhost' IDENTIFIED BY 'testu9' FAILED_LOGIN_ATTEMPTS 1 ")
-	tk.MustQuery("select JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_attempts') " +
-		"from mysql.user where user='testu9' and host ='localhost'").Check(testkit.Rows("1"))
-	tk.MustExec("ALTER USER 'testu9'@'localhost' FAILED_LOGIN_ATTEMPTS 0")
-	tk.MustQuery("select user_attributes from mysql.user where user='testu9' and host ='localhost'").Check(testkit.Rows("{\"resource_group\": \"default\"}"))
-
-	// Specify only PASSWORD_LOCK_TIME one attribute when creating user.
-	// Change the value to 0 and check the user_attributes value.
-	tk.MustExec("CREATE USER 'testu10'@'localhost' IDENTIFIED BY 'testu10' PASSWORD_LOCK_TIME 1 ")
-	tk.MustQuery("select JSON_EXTRACT(user_attributes, '$.Password_locking.password_lock_time_days') " +
-		"from mysql.user where user='testu10' and host ='localhost'").Check(testkit.Rows("1"))
-	tk.MustExec("ALTER USER 'testu10'@'localhost' PASSWORD_LOCK_TIME 0")
-	tk.MustQuery("select user_attributes from mysql.user where user='testu10' and host ='localhost'").Check(testkit.Rows("{\"resource_group\": \"default\"}"))
-
-	// Specify FAILED_LOGIN_ATTEMPTS and PASSWORD_LOCK_TIME attributes when creating user ,
-	// change the values of the two attributes to 0, and check the value of user_attributes.
-	tk.MustExec("CREATE USER 'testu11'@'localhost' IDENTIFIED BY 'testu11' FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 1 ")
-	tk.MustQuery("select JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_attempts')," +
-		"JSON_EXTRACT(user_attributes, '$.Password_locking.password_lock_time_days') " +
-		"from mysql.user where user='testu11' and host ='localhost'").Check(testkit.Rows("1 1"))
-	tk.MustExec("ALTER USER 'testu11'@'localhost' PASSWORD_LOCK_TIME 0")
-	tk.MustQuery("select JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_attempts')," +
-		"JSON_EXTRACT(user_attributes, '$.Password_locking.password_lock_time_days') " +
-		"from mysql.user where user='testu11' and host ='localhost'").Check(testkit.Rows("1 0"))
-	tk.MustExec("ALTER USER 'testu11'@'localhost' FAILED_LOGIN_ATTEMPTS 0")
-	tk.MustQuery("select user_attributes " +
-		"from mysql.user where user='testu11' and host ='localhost'").Check(testkit.Rows("{\"resource_group\": \"default\"}"))
-}
-
-func TestFailedLoginTrackingCheckPrivilges(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	createAndCheck(tk, "CREATE USER 'testu1'@'localhost' IDENTIFIED BY '' FAILED_LOGIN_ATTEMPTS 1 PASSWORD_LOCK_TIME 1",
-		"{\"Password_locking\": {\"failed_login_attempts\": 1, \"password_lock_time_days\": 1}, \"resource_group\": \"default\"}", "testu1")
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: "testu1", Hostname: "localhost"}, nil, nil))
-	// Specify FAILED_LOGIN_ATTEMPTS and PASSWORD_LOCK_TIME attributes when creating user ,
-	// Check user privileges  after successful login.
-	tk.MustQuery(`show grants`).Check(testkit.Rows("GRANT USAGE ON *.* TO 'testu1'@'localhost'"))
-	tk.MustQuery(`select user()`).Check(testkit.Rows("testu1@localhost"))
-}
-
-func TestFailedLoginTracking(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	rootk := testkit.NewTestKit(t, store)
-	createAndCheck(tk, "CREATE USER 'u6'@'localhost' IDENTIFIED BY '' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 3",
-		"{\"Password_locking\": {\"failed_login_attempts\": 3, \"password_lock_time_days\": 3}, \"resource_group\": \"default\"}", "u6")
-	createAndCheck(tk, "CREATE USER 'u5'@'localhost' IDENTIFIED BY '' FAILED_LOGIN_ATTEMPTS 60 PASSWORD_LOCK_TIME 3",
-		"{\"Password_locking\": {\"failed_login_attempts\": 60, \"password_lock_time_days\": 3}, \"resource_group\": \"default\"}", "u5")
-	failedLoginTrackingCase1(t, tk, rootk)
-	failedLoginTrackingCase2(t, tk, rootk)
-	failedLoginTrackingCase3(t, tk, rootk)
-	failedLoginTrackingCase4(t, tk, rootk)
-	failedLoginTrackingCase5(t, tk, rootk)
-	failedLoginTrackingCase6(t, tk, rootk)
-	failedLoginTrackingCase7(t, tk, rootk)
-}
-
-func failedLoginTrackingCase1(t *testing.T, tk *testkit.TestKit, rootk *testkit.TestKit) {
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u6", 1, "N")
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, nil, nil))
-	checkAuthUser(t, rootk, "u6", 0, "N")
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: "root", Hostname: "localhost"}, nil, nil))
-}
-
-func failedLoginTrackingCase2(t *testing.T, tk *testkit.TestKit, rootk *testkit.TestKit) {
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u6", 1, "N")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u6", 2, "N")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u6", 3, "Y")
-}
-
-func failedLoginTrackingCase3(t *testing.T, tk *testkit.TestKit, rootk *testkit.TestKit) {
-	changeAutoLockedLastChanged(rootk, "-72h1s", "u6")
-	loadUser(t, tk, 1, rootk)
-	checkAuthUser(t, rootk, "u6", 3, "Y")
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, nil, nil))
-	checkAuthUser(t, rootk, "u6", 0, "N")
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: "root", Hostname: "localhost"}, nil, nil))
-}
-
-func failedLoginTrackingCase4(t *testing.T, tk *testkit.TestKit, rootk *testkit.TestKit) {
-	failedLoginTrackingCase2(t, tk, rootk)
-	alterAndCheck(t, rootk, "ALTER USER 'u6'@'localhost' ACCOUNT UNLOCK;", "u6", 3, 3, 0)
-	loadUser(t, tk, 2, rootk)
-	checkAuthUser(t, rootk, "u6", 0, "N")
-}
-
-func failedLoginTrackingCase5(t *testing.T, tk *testkit.TestKit, rootk *testkit.TestKit) {
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u6", 1, "N")
-	alterAndCheck(t, rootk, "ALTER USER 'u6'@'localhost' ACCOUNT UNLOCK;", "u6", 3, 3, 0)
-	checkAuthUser(t, rootk, "u6", 0, "N")
-}
-
-func failedLoginTrackingCase6(t *testing.T, tk *testkit.TestKit, rootk *testkit.TestKit) {
-	failedLoginTrackingCase2(t, tk, rootk)
-	changeAutoLockedLastChanged(rootk, "-72h1s", "u6")
-	loadUser(t, tk, 3, rootk)
-	checkAuthUser(t, rootk, "u6", 3, "Y")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u6", 1, "N")
-}
-
-func failedLoginTrackingCase7(t *testing.T, tk *testkit.TestKit, rootk *testkit.TestKit) {
-	createAndCheck(rootk, "CREATE USER 'u1'@'localhost' IDENTIFIED BY '' FAILED_LOGIN_ATTEMPTS 3",
-		"{\"Password_locking\": {\"failed_login_attempts\": 3, \"password_lock_time_days\": 0}, \"resource_group\": \"default\"}", "u1")
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u6", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u1", 0, "")
-	alterAndCheck(t, rootk, "ALTER USER 'u1'@'localhost' PASSWORD_LOCK_TIME 6;", "u1", 3, 6, 0)
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u1", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u1", 1, "N")
-}
-
-func TestUserPassword(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec(`set global validate_password.enable = ON`)
-	runSQLAndCheck(t, store, "CREATE USER 'u1'@'localhost' IDENTIFIED BY 'qwe123' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 4;",
-		"CREATE USER 'u1'@'localhost' IDENTIFIED BY '!@#HASHhs123' FAILED_LOGIN_ATTEMPTS 3 PASSWORD_LOCK_TIME 4;", "u1", "localhost",
-		"{\"Password_locking\": {\"failed_login_attempts\": 3, \"password_lock_time_days\": 4}, \"resource_group\": \"default\"}", "qwe123", "!@#HASHhs123")
-
-	runSQLAndCheck(t, store, `CREATE USER 'u2'@'localhost' IDENTIFIED BY 'qwe123' FAILED_LOGIN_ATTEMPTS 4 PASSWORD_LOCK_TIME 3 COMMENT 'Some statements to test create user'`,
-		`CREATE USER 'u2'@'localhost' IDENTIFIED BY '!@#HASHhs123' FAILED_LOGIN_ATTEMPTS 4 PASSWORD_LOCK_TIME 3 COMMENT 'Some statements to test create user'`, "u2", "localhost",
-		"{\"Password_locking\": {\"failed_login_attempts\": 4, \"password_lock_time_days\": 3}, \"metadata\": {\"comment\": \"Some statements to test create user\"}, \"resource_group\": \"default\"}", "qwe123", "!@#HASHhs123")
-}
-
-func runSQLAndCheck(t *testing.T, store kv.Storage, errSQL, sucSQL, user, host, rsJSON, simplePassword, strongPassword string) {
-	tk := testkit.NewTestKit(t, store)
-	rootk := testkit.NewTestKit(t, store)
-	createAndCheckToErr(t, rootk, errSQL, user)
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: user, Hostname: host}, sha1Password(simplePassword), nil))
-	createAndCheck(rootk, sucSQL,
-		rsJSON, user)
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: user, Hostname: host}, sha1Password(strongPassword), nil))
-}
-
-func TestPasswordExpiredAndTacking(t *testing.T) {
-	store := testkit.CreateMockStore(t)
-	tk := testkit.NewTestKit(t, store)
-	user := "u3"
-	host := "localhost"
-	tk.MustExec(`set global validate_password.enable = ON`)
-	tk = testkit.NewTestKit(t, store)
-	createAndCheckToErr(t, tk, `CREATE USER 'u3'@'localhost' IDENTIFIED BY 'qwe123' PASSWORD EXPIRE INTERVAL 3 DAY FAILED_LOGIN_ATTEMPTS 4 PASSWORD_LOCK_TIME 3 COMMENT 'Some statements to test create user'`, user)
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: user, Hostname: host}, sha1Password("qwe123"), nil))
-	tk = testkit.NewTestKit(t, store)
-	createAndCheck(tk, `CREATE USER 'u3'@'localhost' IDENTIFIED BY '!@#HASHhs123' PASSWORD EXPIRE INTERVAL 3 DAY  FAILED_LOGIN_ATTEMPTS 4 PASSWORD_LOCK_TIME 3 COMMENT 'Some statements to test create user'`,
-		"{\"Password_locking\": {\"failed_login_attempts\": 4, \"password_lock_time_days\": 3}, \"metadata\": {\"comment\": \"Some statements to test create user\"}, \"resource_group\": \"default\"}", user)
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: user, Hostname: host}, sha1Password("!@#HASHhs123"), nil))
-
-	tk = testkit.NewTestKit(t, store)
-	tk.MustExec(fmt.Sprintf("ALTER USER '%s'@'%s' PASSWORD EXPIRE NEVER", user, host))
-	tk = testkit.NewTestKit(t, store)
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: user, Hostname: host}, sha1Password("!@#HASHhs123"), nil))
-
-	loginFailedAncCheck(t, store, user, host, "password", 1, "N")
-	loginSucAncCheck(t, store, user, host, "!@#HASHhs123", 0, "N")
-	loginFailedAncCheck(t, store, user, host, "password", 1, "N")
-	loginFailedAncCheck(t, store, user, host, "password", 2, "N")
-	loginFailedAncCheck(t, store, user, host, "password", 3, "N")
-	loginFailedAncCheck(t, store, user, host, "password", 4, "Y")
-
-	tk = testkit.NewTestKit(t, store)
-	tk.MustExec(fmt.Sprintf("ALTER USER '%s'@'%s' PASSWORD EXPIRE", user, host))
-	tk = testkit.NewTestKit(t, store)
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: user, Hostname: host}, sha1Password("!@#HASHhs123"), nil))
-}
-
-func loginFailedAncCheck(t *testing.T, store kv.Storage, user, host, password string, failedLoginCount int64, autoAccountLocked string) {
-	tk := testkit.NewTestKit(t, store)
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: user, Hostname: host}, sha1Password(password), nil))
-	checkAuthUser(t, tk, user, failedLoginCount, autoAccountLocked)
-}
-
-func loginSucAncCheck(t *testing.T, store kv.Storage, user, host, password string, failedLoginCount int64, autoAccountLocked string) {
-	tk := testkit.NewTestKit(t, store)
-	require.NoError(t, tk.Session().Auth(&auth.UserIdentity{Username: user, Hostname: host}, sha1Password(password), nil))
-	tk = testkit.NewTestKit(t, store)
-	checkAuthUser(t, tk, user, failedLoginCount, autoAccountLocked)
-}
-
-func sha1Password(s string) []byte {
-	crypt := sha1.New()
-	crypt.Write([]byte(s))
-	hashStage1 := crypt.Sum(nil)
-	crypt.Reset()
-	crypt.Write(hashStage1)
-	hashStage2 := crypt.Sum(nil)
-	crypt.Reset()
-	crypt.Write(hashStage2)
-	hashStage3 := crypt.Sum(nil)
-	for i := range hashStage3 {
-		hashStage3[i] ^= hashStage1[i]
-	}
-	return hashStage3
-}
-
-func loadUser(t *testing.T, tk *testkit.TestKit, useCount int64, rootk *testkit.TestKit) {
-	require.Error(t, tk.Session().Auth(&auth.UserIdentity{Username: "u5", Hostname: "localhost"}, sha1Password("password"), nil))
-	checkAuthUser(t, rootk, "u5", useCount, "N")
-}
-
-func changeAutoLockedLastChanged(tk *testkit.TestKit, ds, user string) {
-	SQL := "UPDATE `mysql`.`User` SET user_attributes=json_merge_patch(user_attributes, '{\"Password_locking\": {\"failed_login_attempts\": 3," +
-		"\"password_lock_time_days\": 3,\"auto_account_locked\": \"Y\",\"failed_login_count\": 3,\"auto_locked_last_changed\": \"%s\"}}') " +
-		"WHERE Host='localhost' and User='%s'"
-	d, _ := time.ParseDuration(ds)
-	changeTime := time.Now().Add(d).Format(time.UnixDate)
-	SQL = fmt.Sprintf(SQL, changeTime, user)
-	tk.MustExec(SQL)
-	domain.GetDomain(tk.Session()).NotifyUpdatePrivilege()
-}
-
-func checkUserUserAttributes(tk *testkit.TestKit, user, host, row string) {
-	sqlTemplate := "select JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_attempts')," +
-		"JSON_EXTRACT(user_attributes, '$.Password_locking.auto_account_locked')," +
-		"JSON_EXTRACT(user_attributes, '$.Password_locking.failed_login_count')," +
-		"JSON_EXTRACT(user_attributes, '$.Password_locking.password_lock_time_days')," +
-		"JSON_EXTRACT(user_attributes, '$.metadata')from mysql.user where user= %? and host = %?"
-	userAttributesSQL := new(strings.Builder)
-	sqlexec.MustFormatSQL(userAttributesSQL, sqlTemplate, user, host)
-	tk.MustQuery(userAttributesSQL.String()).Check(testkit.Rows(row))
-}
-
-func alterAndCheck(t *testing.T, tk *testkit.TestKit, sql string, user string, failedLoginAttempts, passwordLockTimeDays, failedLoginCount int64) {
-	tk.MustExec(sql)
-	userAttributesSQL := selectSQL(user)
-	resBuff := bytes.NewBufferString("")
-	rs := tk.MustQuery(userAttributesSQL)
-	for _, row := range rs.Rows() {
-		_, err := fmt.Fprintf(resBuff, "%s\n", row)
-		require.NoError(t, err)
-	}
-	err := checkUser(t, resBuff.String(), failedLoginAttempts, passwordLockTimeDays, failedLoginCount)
-	require.NoError(t, err)
-}
-
-func checkUser(t *testing.T, rs string, failedLoginAttempts, passwordLockTimeDays, failedLoginCount int64) error {
-	var ua []userAttributes
-	if err := json.Unmarshal([]byte(rs), &ua); err != nil {
-		return err
-	}
-	require.True(t, ua[0].PasswordLocking.FailedLoginAttempts == failedLoginAttempts)
-	require.True(t, ua[0].PasswordLocking.PasswordLockTimeDays == passwordLockTimeDays)
-	require.True(t, ua[0].PasswordLocking.FailedLoginCount == failedLoginCount)
-	return nil
-}
-
-func createAndCheck(tk *testkit.TestKit, sql, rsJSON, user string) {
-	tk.MustExec(sql)
-	sql = selectSQL(user)
-	tk.MustQuery(sql).Check(testkit.Rows(rsJSON))
-}
-
-func createAndCheckToErr(t *testing.T, tk *testkit.TestKit, sql, user string) {
-	tk.MustExecToErr(sql)
-	sql = selectSQL(user)
-	require.True(t, len(tk.MustQuery(sql).Rows()) == 0)
-}
-
-func checkAuthUser(t *testing.T, tk *testkit.TestKit, user string, failedLoginCount int64, autoAccountLocked string) {
-	userAttributesSQL := selectSQL(user)
-	resBuff := bytes.NewBufferString("")
-	rs := tk.MustQuery(userAttributesSQL)
-	for _, row := range rs.Rows() {
-		_, err := fmt.Fprintf(resBuff, "%s\n", row)
-		require.NoError(t, err)
-	}
-	var ua []userAttributes
-	err := json.Unmarshal(resBuff.Bytes(), &ua)
-	require.NoError(t, err)
-	require.Equal(t, failedLoginCount, ua[0].PasswordLocking.FailedLoginCount)
-	require.Equal(t, autoAccountLocked, ua[0].PasswordLocking.AutoAccountLocked)
-}
-
-func selectSQL(user string) string {
-	userAttributesSQL := new(strings.Builder)
-	sqlexec.MustFormatSQL(userAttributesSQL, "SELECT user_attributes from mysql.user WHERE USER = %? AND HOST = 'localhost' for update", user)
-	return userAttributesSQL.String()
-}
-
-type userAttributes struct {
-	PasswordLocking passwordLocking `json:"Password_locking"`
-}
-
-type passwordLocking struct {
-	FailedLoginAttempts   int64  `json:"failed_login_attempts"`
-	PasswordLockTimeDays  int64  `json:"password_lock_time_days"`
-	AutoAccountLocked     string `json:"auto_account_locked"`
-	FailedLoginCount      int64  `json:"failed_login_count"`
-	AutoLockedLastChanged string `json:"auto_locked_last_changed"`
 }
